requirements.txt
streamlit==1.34.0
cmake
wheel
pyarrow==14.0.0   
boto3==1.18.0
# pillow==7.2.0
python-docx==0.8.11
python-dotenv
protobuf==3.20.*
pydantic
pinyin
cutlet
azure-ai-vision-imageanalysis
azure-ai-textanalytics==5.2.0